refactor: clean up CLI args, fix naming and parameter types
This commit overhauls command line argument handling and consistent naming across the codebase:
- Rename --chunk CLI flag to --chunk-records for fa window command, update docs and error messages
- Fix argument naming in fa split, fa one, fas multiz, fas refine, pl prefilter commands
- Standardize net subcommand flags to --target-names/--query-names instead of bare --target/--query
- Replace duplicate arg definitions with shared helper functions for in_net/in_chain args
- Update mat subset and twobit some commands to use shared fa name list arg helpers
- Fix axt to fas command to use shared size file arg helper
- Change min_len parameter in PafQueryOptions and BFS code from i32 to usize for proper unsigned interval length handling
- Update all related docs, tests and argument parsing logic

+++ src/libs/paf/query.rs
@@ -29,7 +29,7 @@ pub struct QueryOptions {
     /// Maximum BFS depth (0 = unlimited).
     pub max_depth: u16,
     /// Minimum interval length to propagate.
-    pub min_len: i32,
+    pub min_len: usize,
     /// Minimum distance to merge adjacent intervals.
     pub min_dist: i32,
     /// Minimum gap-compressed identity (0.0-1.0).
